Huge change to allow teams to be associated with espn user
Previously the epsn_user was linked to the league and so
there was a many to many relationship that included the
team_id. This change eliminates the need for that manytomany
relationship by replacing it with the one to many team->user
relation. Because it was a big data change, I started the migrations
over.