[FIX] Update Espn Id On Load Player Game
If a player was created during the load of the
transaction log, it would not get the espn id,
and load_games was not fixing that even though
it’s necessary to calculate optimal lineup.

Made the integration test fail by raising an 
exception when the optimal lineup points is
less than the actual lineup, this gave me
a red text that I fixed.